move component filter result
