+++ components/indicadors/FilterResult.js
